Relocated pytest configuration to pyproject.toml
The pytest configuration previously contained in pytest.ini has been migrated to pyproject.toml in order to go along with the move to use Poetry for dependency management in the project. The pytest configuration in 'my-pytest.ini' remains the same to ensure backward compatibility for those still depending on the .ini files.